requirements.txt
fastapi~=0.114.2
uvicorn
asyncpg
sqlalchemy~=2.0.35
alembic~=1.13.2
psycopg2-binary
pydantic
python-dotenv~=1.0.1
passlib~=1.7.4
python-jose~=3.3.0
aiogram==3.13.0
aiohttp~=3.10.5